+++ src/main/java/fr/alasdiablo/janoeo/arsenal/utils/ArmorsMaterials.java
@@ -1,8 +1,10 @@
 package fr.alasdiablo.janoeo.arsenal.utils;
 
+import fr.alasdiablo.janoeo.ingots.IngotsItems;
 import fr.alasdiablo.janoeo.tags.ItemsTags;
 import net.minecraft.inventory.EquipmentSlotType;
 import net.minecraft.item.IArmorMaterial;
+import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.SoundEvents;
@@ -12,28 +14,28 @@ import net.minecraft.util.SoundEvents;
  */
 public enum ArmorsMaterials implements IArmorMaterial {
 
-    BLACK_WOOL_ARMOR(Registries.BLACK_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    BLUE_WOOL_ARMOR(Registries.BLUE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    BROWN_WOOL_ARMOR(Registries.BROWN_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    CYAN_WOOL_ARMOR(Registries.CYAN_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    GRAY_WOOL_ARMOR(Registries.GRAY_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    GREEN_WOOL_ARMOR(Registries.GREEN_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    LIGHT_BLUE_WOOL_ARMOR(Registries.LIGHT_BLUE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    LIGHT_GRAY_WOOL_ARMOR(Registries.LIGHT_GRAY_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    LIME_WOOL_ARMOR(Registries.LIME_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    MAGENTA_WOOL_ARMOR(Registries.MAGENTA_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    ORANGE_WOOL_ARMOR(Registries.ORANGE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    PINK_WOOL_ARMOR(Registries.PINK_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    PURPLE_WOOL_ARMOR(Registries.PURPLE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    RED_WOOL_ARMOR(Registries.RED_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    WHITE_WOOL_ARMOR(Registries.WHITE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    YELLOW_WOOL_ARMOR(Registries.YELLOW_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f),
-    COPPER_ARMOR(Registries.COPPER_ARMOR,  7, new int[]{1, 3, 5, 2}, 9, Ingredient.fromTag(ItemsTags.Ingots.COPPER), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F),
-    ALUMINIUM_ARMOR(Registries.ALUMINIUM_ARMOR,  6, new int[]{1, 2, 4, 1}, 6, Ingredient.fromTag(ItemsTags.Ingots.ALUMINIUM), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F),
-    LEAD_ARMOR(Registries.LEAD_ARMOR,  5, new int[]{1, 2, 4, 1}, 5, Ingredient.fromTag(ItemsTags.Ingots.LEAD), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F),
-    SILVER_ARMOR(Registries.SILVER_ARMOR,  15, new int[]{2, 5, 6, 2}, 25, Ingredient.fromTag(ItemsTags.Ingots.SILVER), SoundEvents.ITEM_ARMOR_EQUIP_GOLD, 0.0F),
-    TIN_ARMOR(Registries.TIN_ARMOR, 6, new int[]{1, 2, 4, 1}, 7, Ingredient.fromTag(ItemsTags.Ingots.TIN), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F),
-    URANIUM_ARMOR(Registries.URANIUM_ARMOR,  12, new int[]{2, 5, 6, 2}, 9, Ingredient.fromTag(ItemsTags.Ingots.URANIUM), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F);
+    BLACK_WOOL_ARMOR(Registries.BLACK_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    BLUE_WOOL_ARMOR(Registries.BLUE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    BROWN_WOOL_ARMOR(Registries.BROWN_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    CYAN_WOOL_ARMOR(Registries.CYAN_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    GRAY_WOOL_ARMOR(Registries.GRAY_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    GREEN_WOOL_ARMOR(Registries.GREEN_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    LIGHT_BLUE_WOOL_ARMOR(Registries.LIGHT_BLUE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    LIGHT_GRAY_WOOL_ARMOR(Registries.LIGHT_GRAY_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    LIME_WOOL_ARMOR(Registries.LIME_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    MAGENTA_WOOL_ARMOR(Registries.MAGENTA_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    ORANGE_WOOL_ARMOR(Registries.ORANGE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    PINK_WOOL_ARMOR(Registries.PINK_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    PURPLE_WOOL_ARMOR(Registries.PURPLE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    RED_WOOL_ARMOR(Registries.RED_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    WHITE_WOOL_ARMOR(Registries.WHITE_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    YELLOW_WOOL_ARMOR(Registries.YELLOW_WOOL_ARMOR, 5, new int[] {1, 2, 3, 1} ,25, Ingredient.fromItems(Registries.WOOLS_LIST), SoundEvents.ITEM_ARMOR_EQUIP_LEATHER, 0.2f, 0F),
+    COPPER_ARMOR(Registries.COPPER_ARMOR,  7, new int[]{1, 3, 5, 2}, 9, Ingredient.fromStacks(new ItemStack(IngotsItems.COPPER_INGOT)), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F, 0F),
+    ALUMINIUM_ARMOR(Registries.ALUMINIUM_ARMOR,  6, new int[]{1, 2, 4, 1}, 6, Ingredient.fromStacks(new ItemStack(IngotsItems.ALUMINIUM_INGOT)), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F, 0F),
+    LEAD_ARMOR(Registries.LEAD_ARMOR,  5, new int[]{1, 2, 4, 1}, 5, Ingredient.fromStacks(new ItemStack(IngotsItems.LEAD_INGOT)), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F, 0F),
+    SILVER_ARMOR(Registries.SILVER_ARMOR,  15, new int[]{2, 5, 6, 2}, 25, Ingredient.fromStacks(new ItemStack(IngotsItems.SILVER_INGOT)), SoundEvents.ITEM_ARMOR_EQUIP_GOLD, 0.0F, 0F),
+    TIN_ARMOR(Registries.TIN_ARMOR, 6, new int[]{1, 2, 4, 1}, 7, Ingredient.fromStacks(new ItemStack(IngotsItems.TIN_INGOT)), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F, 0F),
+    URANIUM_ARMOR(Registries.URANIUM_ARMOR,  12, new int[]{2, 5, 6, 2}, 9, Ingredient.fromStacks(new ItemStack(IngotsItems.URANIUM_INGOT)), SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0F, 0F);
 
     /**
      * name of the material
@@ -70,6 +72,8 @@ public enum ArmorsMaterials implements IArmorMaterial {
      */
     private float toughness;
 
+    private float knockBackResistance;
+
     /**
      * default constructor
      * @param name name of the material
@@ -80,7 +84,7 @@ public enum ArmorsMaterials implements IArmorMaterial {
      * @param soundEvent sound(sound use on equipe) of the material
      * @param toughness toughness of the material
      */
-    ArmorsMaterials(String name, int durability, int[] damageReductionAmount, int enchantability, Ingredient repairItem, SoundEvent soundEvent, float toughness) {
+    ArmorsMaterials(String name, int durability, int[] damageReductionAmount, int enchantability, Ingredient repairItem, SoundEvent soundEvent, float toughness, float knockBackResistance) {
         final int[] max_damge_arry = new int[] {13, 15, 16, 11};
         this.damageReductionAmount = damageReductionAmount;
         this.toughness = toughness;
@@ -92,6 +96,7 @@ public enum ArmorsMaterials implements IArmorMaterial {
         }
         this.repairMaterial = repairItem;
         this.name = Registries.MOD_ID + ":" + name;
+        this.knockBackResistance = knockBackResistance;
     }
 
     /**
@@ -157,5 +162,10 @@ public enum ArmorsMaterials implements IArmorMaterial {
     @Override
     public float getToughness() {
         return this.toughness;
+    }
+
+    @Override
+    public float func_230304_f_() {
+        return this.knockBackResistance;
     }
 }